interface: moved components/lib to components
