+++ pkg/interface/src/views/apps/chat/components/ChatInput.tsx
