+++ pkg/interface/src/views/apps/chat/components/ChatWindow.tsx
