+++ pkg/interface/src/views/apps/chat/components/content/code.js
