+++ pkg/interface/src/views/apps/chat/components/content/text.js
