+++ pkg/interface/src/views/apps/chat/components/overlay-sigil.js
