+++ pkg/interface/src/views/apps/chat/components/profile-overlay.js
